refactor: 💡 BASE_URL/schedule/lesson/{lessonIdx} url만 변경
